Fixed apparent problems with assigning getnum (which can return -1) to unit.

+++ sbbs3/upload.cpp
@@ -2,7 +2,7 @@
 
 /* Synchronet file upload-related routines */
 
-/* $Id: upload.cpp,v 1.3 2000/10/30 08:49:26 rswindell Exp $ */
+/* $Id: upload.cpp,v 1.4 2000/11/02 23:30:39 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -388,10 +388,10 @@ void sbbs_t::upload(uint dirnum)
 		SYNC;
 		if(!noyes(text[MultipleDiskQ])) {
 			bputs(text[HowManyDisksTotal]);
-			if((i=getnum(99))<2)
+			if((int)(i=getnum(99))<2)
 				return;
 			bputs(text[NumberOfFile]);
-			if((j=getnum(i))<1)
+			if((int)(j=getnum(i))<1)
 				return;
 			if(j==1)
 				upload_lastdesc[0]=0;
